requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
openai==0.27.8 
requests==2.31.0
beautifulsoup4==4.12.2 
python-dotenv==0.18.0
langchain==0.0.263 
chromadb==0.4.5 
pypdf==3.15.1 
tiktoken==0.4.0 
